Fix issues with GetSmellsByDate
Fixes #2

The issue referenced is #2 on GitHub (https://github.com/carlovend/CADOCS_II/issues/2).

The problem was with the incorrect use of the > operand. When users requested the function to get smells by date, the entities array would contain at least 2 entries. However, in the code, this was always ignored due to the condition `if len(entities) > 2`.

Additionally, the format of the date fed to csDetector was changed to the format accepted by csDetector, which is %Y-%m-%d.

Lastly, the GET parameter in the URL request to csDetector was updated. The "date" parameter was replaced with "start" because csDetector does not use "date" but instead uses "start".